requirements.txt
tkcalendar
mysql-connector-python
pymysql
pathlib
tkinter